requirement.txt
Flask==2.2.2
pandas==1.4.2
spacy==3.3.1
python-dateutil==2.8.2
boto3==1.24.28 
